temporal/requirements.txt
temporalio>=1.8.0,<2
httpx>=0.27.0
python-dotenv>=1.0.0
